refactor: group test fixtures by steam interface and request
mirror the request and test layout in fixtures:
tests/Fixtures/Saloon/<Interface>/<Request>/<variant>.json with
shortened variant names (default, empty, private, ...).

also move the stray GetFriendListRequest (and its test) into the
ISteamUser subnamespace to match the rest of the requests.
