+++ tests/Fixtures/Saloon/ISteamUser/GetFriendList/default.json
