+++ tests/Fixtures/Saloon/ISteamUser/GetFriendList/empty.json
