+++ tests/Fixtures/Saloon/ISteamUser/ResolveVanityUrl/not_found.json
