+++ tests/Fixtures/Saloon/ISteamUser/ResolveVanityUrl/success.json
