+++ tests/Fixtures/Saloon/ISteamUserStats/GetUserStatsForGame/empty.json
