+++ tests/Fixtures/Saloon/ISteamUserStats/GetUserStatsForGame/private.json
